requirements.txt
# Core framework
fastapi>=0.115.5
uvicorn[standard]>=0.30.0

# AWS
boto3>=1.42.63
bedrock-agentcore>=1.4.6

# LangChain / LangGraph
langgraph>=0.2.45
langchain>=0.3.7
langchain-aws>=0.2.7
langchain-community>=0.3.7

# Observability
langfuse>=4.0.0

# Stock data
yfinance>=0.2.44

# Auth
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# Config
pydantic>=2.9.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1

# Notebook
jupyter>=1.1.1
requests>=2.32.3